ffi: assign_partition takes num_tiles, divides by compute_rows
The previous commits passed num_col=arg.num_tiles to the FFI verbatim,
treating it as a column count.  Bridge sweep exposed the unit error:
SHIM's create_ctx_on_device packs num_tiles = m_col_cnt * core_rows
(xdna-driver src/shim/hwctx.cpp:279), so num_tiles=8 means 2 cols on
NPU1 (4 compute rows), not 8.  The naive read tried to ungate cols
1..9 and either failed the bounds check (first attempt) or over-
ungated to the array end (clip attempt).

This invert the formula in the FFI: num_col = num_tiles / compute_rows,
where compute_rows = total_rows - 2 (AIE2: 1 shim + 1 memtile + rest
compute).  Non-multiples reject as malformed.  Out-of-range now means
"upstream sent a partition that doesn't fit", which is a real error
worth surfacing, not a clip-and-continue case.

Plugin C++: assign_partition signature widens to (u8 start_col,
u32 num_tiles); platform_emu.cpp passes ctx_num_tiles (the misnamed
m_ctx_map.num_col field that actually stores num_tiles) directly.
Renaming the field is deferred.

Bridge sweep result with the fix:
  Chess: 0/65 EMU pass -> 57/65 (88%)
  Peano: 0/56 EMU pass -> 50/56 (89%)
  Trace divergence: 0 across both compilers
Remaining failures cluster in dynamic_object_fifo/* and
objectfifo_repeat/* with DMA-lock-acquire wedges -- pre-existing
object-fifo bugs unrelated to clock control, no "access to gated tile"
warnings.

Generated using Claude Code.

+++ xrt-plugin/src/transport.h
@@ -52,8 +52,13 @@ public:
     /// submit because `reset_context` wipes clock state.  No-op if the
     /// backing FFI symbol is not available (older emulator builds --
     /// tests will wedge because nothing else ungates the columns).
+    ///
+    /// `num_tiles` is the XRT-native unit (cols * compute_rows).  SHIM
+    /// packs it as `m_col_cnt * core_rows` in
+    /// xdna-driver `src/shim/hwctx.cpp:279`; the emulator inverts the
+    /// formula in the FFI using its known topology.
     virtual void assign_partition(uint8_t /*start_col*/,
-                                  uint8_t /*num_col*/) {}
+                                  uint32_t /*num_tiles*/) {}
 
     // -- Buffer management ---------------------------------------------------
